Refactor common methods to superclass
